Add `ceil` and move docs
