+++ typed_floats_macros/src/lib.rs
@@ -16,6 +16,9 @@ use impl_self::*;
 mod impl_self_rhs;
 use impl_self_rhs::*;
 
+mod add_doc;
+use add_doc::*;
+
 fn get_specifications() -> Vec<(&'static str, FloatSpecifications)> {
     vec![
         (
@@ -149,13 +152,7 @@ pub fn generate_floats(_input: proc_macro::TokenStream) -> proc_macro::TokenStre
 
     let mut output = proc_macro2::TokenStream::new();
 
-    output.extend(comment_line(
-        "/// When the result is [`f64`], it may be `NaN`.",
-    ));
-    output.extend(generate_op_table(&floats_f64, "+"));
-    output.extend(generate_op_table(&floats_f64, "-"));
-    output.extend(generate_op_table(&floats_f64, "%"));
-    output.extend(generate_op_table(&floats_f64, "/"));
+    output.extend(generate_main_description(&floats_f64));
     output.extend(quote! {
         pub trait Float {}
     });
@@ -187,6 +184,7 @@ pub fn generate_tests(_input: proc_macro::TokenStream) -> proc_macro::TokenStrea
                         println!("num={:?}", num_a);
                         println!("neg={:?}", -num_a);
                         println!("floor={:?}", num_a.floor());
+                        println!("ceil={:?}", num_a.ceil());
                         println!("abs={:?}", num_a.abs());
                     }
                     Err(_) => {}
@@ -247,59 +245,6 @@ pub fn generate_tests(_input: proc_macro::TokenStream) -> proc_macro::TokenStrea
         }
     }
     .into()
-}
-
-fn comment_line(str: &str) -> proc_macro2::TokenStream {
-    str.parse().unwrap()
-}
-
-fn generate_op_table(floats: &[FloatDefinition], op: &str) -> proc_macro2::TokenStream {
-    let mut output = proc_macro2::TokenStream::new();
-
-    let mut str: String = format!("/// |  {op}  |");
-    for rhs in floats {
-        str += format!(" {rhs_name} |", rhs_name = rhs.name).as_str();
-    }
-    str += "\n";
-
-    output.extend(comment_line(&str));
-
-    let mut str: String = "/// |-|".to_string();
-    for _ in floats {
-        str += "-|";
-    }
-    str += "\n";
-
-    output.extend(comment_line(&str));
-
-    for float in floats {
-        let name = float.name;
-        let float_type = float.float_type;
-
-        let mut str: String = format!("/// {name} | ").to_string();
-
-        for rhs in floats {
-            let result = match op {
-                "+" => add_result(&float.s, &rhs.s, floats),
-                "-" => sub_result(&float.s, &rhs.s, floats),
-                "%" => rem_result(&float.s, &rhs.s, floats),
-                "/" => div_result(&float.s, &rhs.s, floats),
-                _ => panic!("Unknown op {}", op),
-            };
-
-            let result_str = match result {
-                Some(result) => result.name,
-                None => float_type,
-            };
-            str += format!(" {result_str} |").as_str();
-        }
-
-        output.extend(comment_line(&str));
-    }
-
-    output.extend(comment_line("///\n"));
-
-    output
 }
 
 fn do_generate_floats(floats: &[FloatDefinition], with_generic: bool) -> proc_macro2::TokenStream {
@@ -406,6 +351,7 @@ fn do_generate_floats(floats: &[FloatDefinition], with_generic: bool) -> proc_ma
     for float_a in floats {
         output.extend(impl_neg(float_a, floats));
         output.extend(impl_floor(float_a, floats));
+        output.extend(impl_ceil(float_a, floats));
         output.extend(impl_abs(float_a, floats));
 
         for float_b in floats {
